Refactor v2 collection sort helpers
